extract thread config into bitnet-thread-config-core
