+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
   "crates/bitnet-math",
   "crates/bitnet-simd",
   "crates/bitnet-warn-once",
+  "crates/bitnet-thread-config-core", # SRP: reusable thread config and work partitioning primitives
   "crates/bitnet-runtime-context",
   "crates/bitnet-runtime-context-core",
   "crates/bitnet-bdd-grid",
@@ -131,6 +132,7 @@ default-members = [
   "crates/bitnet-math",
   "crates/bitnet-simd",
   "crates/bitnet-warn-once",
+  "crates/bitnet-thread-config-core",
   "crates/bitnet-quantization",
   "crates/bitnet-quantization-bits",
   "crates/bitnet-rope",
@@ -412,6 +414,7 @@ harness = false
 bitnet-math = { path = "crates/bitnet-math", version = "0.2.1-dev" }
 bitnet-quantization-bits = { path = "crates/bitnet-quantization-bits", version = "0.2.1-dev" }
 bitnet-warn-once = { path = "crates/bitnet-warn-once", version = "0.2.1-dev" }
+bitnet-thread-config-core = { path = "crates/bitnet-thread-config-core", version = "0.2.1-dev" }
 bitnet-atomic-file-core = { path = "crates/bitnet-atomic-file-core", version = "0.2.1-dev" }
 # Core dependencies
 anyhow = "1.0.100"
