+++ crates/bitnet-thread-config-core/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "bitnet-thread-config-core"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+authors.workspace = true
+keywords.workspace = true
+categories.workspace = true
+description = "SRP thread configuration and work partitioning primitives"
+
+[dependencies]
